Resolve "integration unit tests"

Closes #17

See merge request leardev/diablo/diablosharp!18

+++ src/DiabloSharp/Models/OAuthToken.cs
@@ -12,6 +12,6 @@ namespace DiabloSharp.Models
         public string Type { get; set; }
 
         [DataMember(Name = "expires_in")]
-        public int SecondsUntilExpiration { get; set; }
+        public long SecondsUntilExpiration { get; set; }
     }
 }
